fix Branch & Customer

+++ PawnShopBE/PawnShopBE.Core/Models/Branch.cs
@@ -15,13 +15,12 @@ namespace PawnShopBE.Core.Models
         public string BranchName { get; set; }
         public string Address { get; set; }
         public string PhoneNumber { get; set; }
-        public int Fund { get; set; }
+        public decimal Fund { get; set; }
         public DateTime CreateDate { get; set; }
         public DateTime? UpdateDate { get; set; }
         public int Status { get; set; }
 
         //relationship
-
         public ICollection<User>? Users { get; set; }
         public ICollection<Contract>? Contracts { get; set; }
         public ICollection<Ledger>? Ledgers { get; set; }
